backend/requirements.txt
Django==5.0.4
channels==4.1.0
channels-redis==4.2.0
strawberry-graphql-django==0.47.0
strawberry-graphql[channels]==0.236.0
psycopg[binary,pool]==3.1.19
redis==5.0.4
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
google-generativeai==0.5.4
aiortc==1.9.0
daphne==4.1.2
django-cors-headers==4.3.1
pydantic-settings==2.2.1
pillow==10.3.0
ai-edge-litert
numpy>=2.0.0
uvicorn[standard]==0.29.0
python-decouple==3.8
